GlobalActions: Bring back old goodies [2/2]
See https://gerrit.paranoidandroid.co/#/c/45/

This commit adds a setting in Development settings for including
options in the power menu for rebooting into recovery or bootloader,
defauled to off.

When enabled, the Advanced reboot options will only be available once
the device is unlocked or on non-secure lockscreen.

Signed-off-by: @BigBrother1984 <carlosavignano@aospa.co>
Change-Id: Ib77f0be178d15131b99adbe0abb6ba18b2b7c0b6

Conflicts:

	res/values/pa_strings.xml
	res/xml/development_prefs.xml
	src/com/android/settings/DevelopmentSettings.java