Settings: increase default regulatory info drawable dimensions
BitmapDrawable.getIntrinsicWidth/Height returns the bitmap density
scaled to the display density, not the actual bitmap height/width.
The default drawable is 1x1, but on xxhdpi devices, the resolution
is 4x4.  This causes the regulatory information text to never be
shown on xxhdpi devices even when the png has not been overlayed.
Increase the dimension limit to a reasonable 32x32 image.

Change-Id: Id6429d46f8612352d7930fdad4a6d125e7939ea2